Only allocate String when we encounter escape sequence
TODO: It'd be better to not unescape the string at all and instead just
`yield EscapedString` and `EscapedKey` at the end. This would make the
`JSON.Parser` not having to allocate anything (except in case of errors)
and make streaming JSON applications much faster as you do not have to
unescape and escape again.

TODO: Add test cases with escaped strings / keys.